+++ resources/js/app/pages/admin/tickets/details/contents/files/sections/customer-tickets-receipt-model.jsx
@@ -86,10 +86,12 @@ const CustomerTicketsReceiptModel = () => {
     
     
     async function deleteFileImage(id, ticket_id) {
+        if (confirm('Are you sure you wanna delete the image?')) {
         setIsLoading(true)
         await store.dispatch(delete_upload_ticket_files_thunk(id, ticket_id))
         setIsLoading(false)
         handleCancel()
+        }
     }
